backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
websockets>=12.0
PySide6>=6.6.0
mac-vendor-lookup>=0.1.0
openpyxl>=3.0.0
alembic>=1.13.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
